Remove thumbURL delete logic

+++ app/Http/Controllers/Api/ComponentTypeController.php
@@ -18,18 +18,18 @@ class ComponentTypeController extends Controller
      */
     public function index()
     {
-        
+
         try {
             $types = ComponentType::all();
-            return response()->json($types,200);
+            return response()->json($types, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
-    
+
 
     /**
      * Store a newly created resource in storage.
@@ -54,12 +54,11 @@ class ComponentTypeController extends Controller
 
             $type = new ComponentType($data);
             $type->save();
-            return response()->json($type,200);
-
+            return response()->json($type, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
@@ -71,26 +70,21 @@ class ComponentTypeController extends Controller
      */
     public function show($id)
     {
-        try 
-        {
+        try {
             $type = ComponentType::find($id);
-            if($type!=null)
-            {
-                return response()->json($type,200);
+            if ($type != null) {
+                return response()->json($type, 200);
+            } else {
+                return response()->json(["message" => "Type is not found!"], 404);
             }
-            else
-            {
-                return response()->json(["message"=>"Type is not found!"],404);
-            }
-        }
-        catch (\Exception $ex) {
+        } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
-    
+
 
     /**
      * Update the specified resource in storage.
@@ -111,8 +105,8 @@ class ComponentTypeController extends Controller
 
         try {
             $componentType = ComponentType::find($id);
-            if($componentType==null){
-                return response()->json(["message"=>"Item not found!"],404);
+            if ($componentType == null) {
+                return response()->json(["message" => "Item not found!"], 404);
             }
 
             if ($request->thumb != null) {
@@ -120,12 +114,11 @@ class ComponentTypeController extends Controller
             }
 
             $componentType->update($data);
-            return response()->json($componentType,200);
-
+            return response()->json($componentType, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
@@ -139,28 +132,27 @@ class ComponentTypeController extends Controller
     {
         try {
             $componentType = ComponentType::find($id);
-            if($componentType==null){
+            if ($componentType == null) {
                 return response()->json([
-                    "message"=>"Item is not found"
-                ],404);
+                    "message" => "Item is not found"
+                ], 404);
             }
             // Delete the thumbnail form the file system
             $this->deleteThumb($componentType->thumbURL());
 
             $componentType->delete();
-            return response()->json($componentType,200);
-
+            return response()->json($componentType, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
 
     private function deleteThumb($currentURL)
     {
-        if ($currentURL != null) {
+        if ($currentURL != null && $currentURL != config('constants.frontend.dummy_thumb')) {
             $oldImage = public_path($currentURL);
             if (File::exists($oldImage)) unlink($oldImage);
         }
@@ -181,19 +173,19 @@ class ComponentTypeController extends Controller
         return $imageName;
     }
 
-    public function search(Request $request){
+    public function search(Request $request)
+    {
         // Get the search value from the request
         $term = $request->query('term');
-    
+
         // search in the title and body columns from the posts table
         $types = ComponentType::query()
             ->where('title', 'LIKE', "%{$term}%")
             ->orWhere('subtitle', 'LIKE', "%{$term}%")
             ->orWhere('description', 'LIKE', "%{$term}%")
             ->get();
-    
+
         // Return the search view with the resluts compacted
         return $types;
     }
-    
 }
