+++ app/Http/Controllers/Api/EquipmentTypeController.php
@@ -20,15 +20,15 @@ class EquipmentTypeController extends Controller
     {
         $types = EquipmentType::all();
         try {
-            return response()->json($types,200);
+            return response()->json($types, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
-    
+
 
     /**
      * Store a newly created resource in storage.
@@ -53,12 +53,11 @@ class EquipmentTypeController extends Controller
 
             $type = new EquipmentType($data);
             $type->save();
-            return response()->json($type,200);
-
+            return response()->json($type, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
@@ -70,26 +69,21 @@ class EquipmentTypeController extends Controller
      */
     public function show($id)
     {
-        try 
-        {
+        try {
             $type = EquipmentType::find($id);
-            if($type!=null)
-            {
-                return response()->json($type,200);
+            if ($type != null) {
+                return response()->json($type, 200);
+            } else {
+                return response()->json(["message" => "Type is not found!"], 404);
             }
-            else
-            {
-                return response()->json(["message"=>"Type is not found!"],404);
-            }
-        }
-        catch (\Exception $ex) {
+        } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
-    
+
 
     /**
      * Update the specified resource in storage.
@@ -109,11 +103,11 @@ class EquipmentTypeController extends Controller
         ]);
 
         try {
-            
+
 
             $equipmentType  = EquipmentType::find($id);
-            if($equipmentType ==null){
-                return response()->json(["message"=>"Item not found!"],404);
+            if ($equipmentType == null) {
+                return response()->json(["message" => "Item not found!"], 404);
             }
             if ($request->thumb != null) {
                 $data['thumb'] = $this->uploadThumb($equipmentType->thumbURL(), $request->thumb, "equipment_items");
@@ -121,12 +115,11 @@ class EquipmentTypeController extends Controller
 
 
             $equipmentType->update($data);
-            return response()->json($equipmentType,200);
-
+            return response()->json($equipmentType, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
@@ -140,27 +133,26 @@ class EquipmentTypeController extends Controller
     {
         try {
             $equipmentType = EquipmentType::find($id);
-            if($equipmentType==null){
+            if ($equipmentType == null) {
                 return response()->json([
-                    "message"=>"Type is not found"
-                ],404);
+                    "message" => "Type is not found"
+                ], 404);
             }
             // Delete the thumbnail form the file system
             $this->deleteThumb($equipmentType->thumbURL());
 
             $equipmentType->delete();
-            return response()->json($equipmentType,200);
-
+            return response()->json($equipmentType, 200);
         } catch (\Exception $ex) {
             return response()->json([
-                "message"=>$ex->getMessage()
-            ],500);
+                "message" => $ex->getMessage()
+            ], 500);
         }
     }
 
     private function deleteThumb($currentURL)
     {
-        if ($currentURL != null) {
+        if ($currentURL != null && $currentURL != config('constants.frontend.dummy_thumb')) {
             $oldImage = public_path($currentURL);
             if (File::exists($oldImage)) unlink($oldImage);
         }
@@ -181,17 +173,18 @@ class EquipmentTypeController extends Controller
         return $imageName;
     }
 
-    public function search(Request $request){
+    public function search(Request $request)
+    {
         // Get the search value from the request
         $term = $request->query('term');
-    
+
         // search in the title and body columns from the posts table
         $types = EquipmentType::query()
             ->where('title', 'LIKE', "%{$term}%")
             ->orWhere('subtitle', 'LIKE', "%{$term}%")
             ->orWhere('description', 'LIKE', "%{$term}%")
             ->get();
-    
+
         // Return the search view with the resluts compacted
         return $types;
     }
